bridge & tainted-ref: only inspect our code
The dist file of bridge & tainted-ref are the biggest contributors[0]
to "BUG"s reported in sonarcloud - but the information is neither
relevant nor actionable. Excluding these accordingly.

node_modules are automatically "excluded by the scm ignore settings"
as could be seen in the logs before, but this seems to happen (somewhat
rightfully so) on individual file basis and consequently is extremely
slow (~40sec on CI). Excluding the whole folder manually.

Having read [1] I'm not convinced that exclusions for the tests are
correct/needed but mended them for bridge & tainted-ref according to how
it was done before.

[0] https://sonarcloud.io/project/issues?id=mediawiki-extensions-Wikibase&resolved=false&types=BUG
[1] https://docs.sonarqube.org/latest/project-administration/narrowing-the-focus/

Change-Id: Ib6a6802c85b01e2b0a82a1052150fb42fb9e9240