Stop locking and use DISTINCT when finding used terms to delete
Right now, it locks millions of rows everywhere just to find if the
locked terms are used. This is unacceptable and causes deadlock
everywhere. Just using master should be enough to make sure we clean up
the unused rows. Worst case, we end up with a negligible number of unused rows

Here's an example of an error in logstash: https://w.wiki/AzE

We lock the rows if we find cases to delete from the new term store
and only lock rows related to those. This might cause unused records
when a row is being used twice, and both gets dropped at the same time.

Bug: T234948
Change-Id: I0e62511d25c1c35daf0eaa0b191b775fbb1cca83