bridge: change to generic service container implementation
Core benefit, aside of a taking a sizable chunk out of our dist code,
is that `ServiceContainer` is now no longer a series of copies of the
same two methods, to be replicated (and tested) over and over.
We entrust these tasks to the compiler and runtime.
I refrained from experiments to further automate the unit tests for
the `ServiceContainer` device (e.g automagically iterating over keys of
`Services`) for now.

Change-Id: I31bbc092868b5a3b4f5f4e64632f4998921bb288