Retrying "Stop locking and use DISTINCT when finding used terms to delete"
This reverts commit 20bfdb321955bf0bf28f57fa77a6ca398b461681
with some changes.

Locking causes long deadlocks and if we stop doing that, the query
actually tries to read millions of rows, this is a middle-ground
solution that fixes the issue of locking while not reading millions
of rows.

The queries have been tested in production and they work as expected.

Bug: T234948
Change-Id: I833083140d657cb48c34c25bc20f6d85168e8690