bridge: prepare e2e tests for entity info requests
The introduction of DispatchingPropertyDataTypeRepository using the
generic ForeignApiEntityInfoDispatcher from the top level of the
application will make it necessary to always properly mock its response
(i.e. it has to match the link we 'click').
This dependency between link and mocked repo is now also reflected in
the test setup (moving their creation into one place).
I agree it is not particularly elegant and welcome any suggestion.

Bug: T236236
Change-Id: I51466fb1d736b01c816cdb0625ad1a04f3db2f8e