bridge: Add reference section component
The ReferenceSection component is intended to own, i.e. encapsulate and
orchestrate, the handling of references. This includes the handling of
the store interactions, which, although there is currently only one,
will become more as we add proper rendering for the Snaks and especially
when we add reference editing.

This implementation uses a custom message for the Snak separator. That
way, individual wikis can overwrite its value without breaking other
places where it is used.

Based on the mock[0], the highlighting of the reference items is intended
to start somewhat left to general center column, but the headline and
the reference text itself should still align with the other headlines in
the center column. This requires some small calculation to ensure that
this works out reliably and that calculation should also make the intend
clearer to future devs.
Also, the horizontal margins and paddings are intended to not scale with
the font size, therefore they are given as px values.

The styling of the "References" heading is intended to be the same as
the Label above in terms of font-size, boldness, etc.

[0]: https://www.figma.com/file/KIkcQ7NjOl8OkQtGWe5duIbW/Wikidata-Bridge?node-id=718%3A1

Bug: T237023
Change-Id: I87f6e098c5477baa372e97926139b82621221a67