TR: mount popper in App.vue
Mount the popper depending on the value in the store.

Wrap both the icon and the Popper in an additional div to ensure they are
both not visible if the statement is untainted.

Add some extra divs to the HTML and style appropriately in an attempt
to get the layout right.

Bug: T237734
Change-Id: I498a8a38031b8951b0edd467b8ae0a4c9cfae718