bridge: Fix margin of Input component
This now aligns with the mocks[0] and, after talking this through with
UX, no longer scales with the font-size.
The reasoning behind this is that growing horizontal paddings lead to
even less horizontal space for a larger font and that doesn't benefit
the reader.

It extracts the margin to the left into a variable so it can be reused
in other components that may need to use it in calculations.

[0]: https://www.figma.com/file/KIkcQ7NjOl8OkQtGWe5duIbW/Wikidata-Bridge?node-id=1526%3A162

Change-Id: Ide3e00d0401eb2006dce718c8988fc054d6e6258