Add ChangeOpResult::validate
This will allow executing validations after applying ChangeOp to an entity
and getting back the ChangeOpResult, which enables further optimizations
on store-wide (across entities) expensive validations by running them only
on the parts that are going to be changed on the entity.

Example under scope atm is implementing Fingerprint uniqueness validations
for Items and Properties in new store.

Bug: T232040
Change-Id: I39cf2323eeeaf804f2c0471267ece1b920a2b134