bridge: attach link listeners as soon as possible
Instead of waiting for ResourceLoader to load the app module before
attaching event listeners to the edit links, attach the listeners
immediately and wait for ResourceLoader within them. This reduces the
time during which clicking the edit links will send the user to the link
target (the repo wiki) instead of opening the data bridge.

To give the tests some chance to figure out when it’s safe to click the
links, the init module still waits for ResourceLoader at the end, so
that when the init module’s promise resolves, the click handler is known
to await an already resolved promise. However, awaiting a resolved
promise still delays execution until the end of the main stack, so we
need to add a call to budge() (i. e., setTimeout(…, 0)) after clicking
the link in several tests.

We also now need to watch out for the case that a user clicks the bridge
link multiple times while we’re waiting for ResourceLoader; introduce an
auxiliary variable to track this case and ignore the extra click events
until we’ve launched the dispatcher, which will cover the whole page in
OOUI’s window manager and prevent further clicks. (It is still possible
to get in an extra click during a very brief window after we’ve called
the dispatcher but before the window manage is fully initialized,
thereby launching a second copy of the app, but that was already the
case before and should be fixed separately.)

On my system, the time between the beginning of the init function and
the click handlers being attached went down from 280±34 ms to 8±2 ms.

Bug: T235765
Change-Id: I3a6d891cdb582099b20f4fd862fbecb93d79ad16