Add ChangeOpFingerprint and ChangeOpFingerprintResult classes
These are decorators for ChangeOps objects (unfortunately using
inheritance for now as ChangeOps does not have an interface).
They can be used for grouping change ops, and later identifying
that group's result, of Fingerprint changes on an entity.

Example use-case atm is fingerprint uniqueness validation that
will need access to all ChangeOpResult instances that are
related to fingerprint parts. In a later patch, these two classes
will be receiving TermValidatorFactory, which in turn will allow
creating uniqueness validator(s) that are based on new store.

Bug: T232040
Change-Id: Ibef6909fbd44205a3c05aa7ebe2cf7fe8bf5d89b