bridge: track bridge opening performance
A different tracker "service" is created in order to keep the separation
between init and app alive and to not have them depend on each others
interfaces unnecessarily. However, the event tracker interface has been
expanded to allow for timing related tracking, because that is likely
going to be helpful in the bridge app itself as well.

Bug: T237368
Change-Id: Ie484e1e1a671b180cf220db168fa60802a2437fc