Don’t explicitly configure stubs in Phan config
Time was, the MediaWiki Phan config did not include the .phan/stubs/
directory by default, and in our config, each of the stubs was guarded
by a check whether the corresponding extension was present (in which
case, it was reasoned, the stub was not necessary and should not be
included). It is different now: those conditionals were removed in
change I984e511d60 (commit cb05500627), so that the stubs are now always
loaded, and since change Icacb42613, the upstream config also includes
.phan/stubs/ by default. So there’s no need to configure it here at all,
and we can simply remove all the lines that mention this directory.

Change-Id: I4668495bcbc84fc1f057967e3bc1f36c422e4581