TR: Close the popper on focusout
And make sure that clicking on the icon also closes
it (and doesn't immediately re-opens it), but making
the icon un-clickable if the popper is open.

Note: Just making the icon on click handler conditional
wont suffice as that will get re-enabled after the
popper closed, but before the very same click event
got handled, thus we need an entirely new element.

Bug: T236889
Change-Id: I5dad7680a7e0a48be1e518d73c28812b7b819fb4