Fix code style issues in PHP entry points
Since we want to migrate to extension registration and get rid of the
PHP entry points in the long run (T88258), all of this code will
eventually have to move somewhere else, where it will be subject to the
usual phpcs code style requirements; those moves will be easier to do
and understand if we already fix the code style issues now, so that the
moves can be as clean as possible.

Most of the fixes are to use imports instead of fully qualified names;
there were also a few other issues, such as using require like a
function, or using the old array() syntax.

Change-Id: I3feb1794a3c542c627232719f30e5da5678c799f