Fix numeric site IDs in dispatch coordinator
For numeric-looking keys, SqlChangeDispatchCoordinator::initState()
would send integer site IDs to the database, which (depending on
MediaWiki rdbms lib version and database implementation) might actually
be stored as integers, which meant that they could not be found in
lockClient() when selecting rows with a certain *string* site ID. Fix
this by casting the site ID to string before inserting it.

This was uncovered by LockManagerSqlChangeDispatchCoordinatorTest, but
that’s not really that test’s responsibility, so instead that test is
changed to specify a more conventional site ID, and a dedicated test for
numeric-looking IDs is added to SqlChangeDispatchCoordinatorTest.

Bug: T238575
Change-Id: I581935a916dd106414922837891ef083d9293e93