Add statementTracker for tracking statement changes
Add statementTracker, for tracking several use cases of statement editing.
It is invoked in init.ts by the save hook.

Additional type definitions for the wikibase-data-model are also added.

Internally statementTracker is using a newly introduced reference
list change counter. Tests for this are to be included as a follow-up
because comprehensively testing require including wikibase-data-model
as a dev dependency.

Bug: T238401
Change-Id: I0e77a867b3cc5843ccca14014b6e84e846192eac