Unsuppress PhanUndeclaredMethod
Most of the issues are fixed with @phan-var, or by suppressing the
issue. The general rule is to suppress all issues coming from the lack
of intersection types support. Phan may support them in the future (https://github.com/phan/phan/issues/1629), so the suppression will be removable at that time.

Change-Id: Ic73bec1c6bd921f6d69e9e8221dec80ebb449ffe