Initializing: animate loading
Modify Initializing to become a decorating component for its
asynchronously loaded content.

Draws inspiration from an article Tonina mentioned[0] while keeping the
contents, which are shown once successfully initialized, dynamic - so
this component can be used time and again, wrapping other things to
load.

The return type of setTimeout() is different in node[1] and window[2]
but, of course, is compatible with its respective clearTimeout()
implementation. TS seems to assume the node version and we don't want to
specify a type (NodeJS.Timer) we feel is "wrong" in browser land, so
let's use a derivative of setTimeout()'s return value[3].

This uses [4] to make testing this (possibly) quicker, resilient, and
a bit more human friedly.

[0] https://alligator.io/vuejs/vue-async-components-load-error/
[1] https://nodejs.org/api/timers.html#timers_settimeout_callback_delay_args
[2] https://www.w3schools.com/jsref/met_win_settimeout.asp
[3] https://stackoverflow.com/a/55607969
[4] https://jestjs.io/docs/en/timer-mocks#advance-timers-by-time

Bug: T237513
Change-Id: I68727f07812884ec372cf24a9d6a939121fd8f21