Use ChangeOpFingerprint to wrap fingerprint (terms) change operations
Excluded from this change: ChangeOpsMerge. This will be done in a
separate follow up patch.

Note for future refactoring:
This is now calling for a ChangeOpBuilder sort of service, to hide the
details of constructing the final tree structure from the deserializers,
for a cleaner and more standard pattern.
That might be even more work, so leaving it out of this task for now.

Bug: T232040
Change-Id: Ia6834d90c5fc4b1b52ac8179a9dee60800560bba