bridge: add messages for users being blocked on the client
These messages can be used to indicate to users that they’re blocked on
the client wiki. The parameters are similar to MediaWiki core’s
blockedtext message, though $4 is not unused in our case: since $1 will
be a DOM element (or jQuery wrapper) containing a link to the user,
which can only be inserted into the message once, $4 is needed to
mention the user again.

Bug: T237526
Change-Id: I230e658412c9d4c9d9fe8bd374d10699180165a4