bridge: add Api, InstantApi, BatchingApi
The BatchingApi class can be used to combine multiple action API
requests into one, as long as they are compatible. It implements a new
Api interface (not to be confused with MediaWiki’s interface, which we
call MwApi), and InstantApi is a simpler implementation of it that can
be used in some tests.

Change-Id: I180a4a59b1c4cc222ef29b75a00705c5a5f33d74