bridge: port two classes to Api
ApiRepoConfigRepository and ApiPageEditPermissionErrorsRepository can
now share the same underlying BatchingApi and have their requests merged
(though ApiRepoConfigRepository needs to change its errorformat for
that; in practice, it can deal with any format other than 'bc', the
default, so 'raw' is fine too).

Some useful definitions from ApiPageEditPermissionErrorsRepository.ts
that aren’t specific to that class are moved to an ApiQuery.ts file.

To simplify the tests, the mockMwApi function is renamed back to mockApi
and declared to return both an Api and an MwApi; since it ignores the
parameters that the function is called with, the interfaces are
sufficiently compatible (“get return a promise”) that this works.

Change-Id: I79f4e1ce102cd6a1093f02f1353dd322619ae4b8