Remove no-op 'jquery.ui.core.styles' from FullEntityParserOutputGenerator
This is a deprecated module alias for 'jquery.ui', however this is only
supported for JS modules. It does not do anything in the styles queue,
except log the following to Logstash:

> resourceloader [ERROR] Unexpected general module "jquery.ui.core.styles"
>                        in styles queue.

To actually work, this will need to load it by other means, which it seems
to be doing already via the JS queue. It does mean that it doesn't load
when JS is disabled, but 1) Afaik we don't render jquery UI buttons from PHP,
and 2) if we do, it will still work correctly as <button> with default
styles.

Bug: T219604
Bug: T239594
Change-Id: Ic0024c32dd14684aa90672d5cd16716892e314fd
(cherry picked from commit 64cd4f014668ca6326acc900d4ee05a1cb0f6fb3)