Allow 'property' param in wbparsevalue to determine datatype
Since a snak's JSON representation includes property ids, but not
the property's datatype. In order to parse/validate input, it'd
be massively useful to not have to lookup the property datatype
via other means but just provide it straight out of the box from
this API - similar to how wbformatvalue already does.

Bug: T239475
Change-Id: Ib7fedbd1447140cc7f118916af3805e02bbd902e