bridge: rename Api interface to MwApi
This frees up the Api name for a different interface of our own.
(Strictly speaking, we don’t need a rename for that, since you can
assign different names on import – but that would be confusing.)

Accompanying types and a few functions are also renamed, but most
variables that contain an instance of MwApi are still called api, not
mwApi – they’ll be renamed on a case-by-case basis, where conflicts with
the new Api interface arise, not all at once.

Change-Id: I422d66d9900aa3738a41aa6b86761f476175db05