Add ByTypeDispatchingEntityIdLookup, allowing other entity id lookup for other types
This removes the existing getEntityIdLookup from ClientStore/DirectSqlStore,
which was always PagePropsEntityIdLookup. Instead, a ByTypeDispatchingEntityIdLookup
will be returned (from WikibaseClient directly) that defaults to this same
original PagePropsEntityIdLookup, but allows for other lookups to be registered
under 'entity-id-lookup-callback'.

WikibaseLanguageIndependentLuaBindings::getEntityId has been changed slightly
to first use said EntityIdLookup when possible, before it goes to use SiteLinkLookup,
which might not be in use for other entity types (e.g. mediainfo)

Bug: T236691
Change-Id: I660ba0aa39314c205c831e9d4714f227575cbd42