bridge: fix InstantApi error handling
The error handling originally implemented in InstantApi was closely
based on existing code in other services, which assumed that the promise
returned by mw.Api rejects with the jqXHR object (and that any errors in
the API response can be found in the resolved value); but as far as I
can tell, that’s inaccurate: the first rejection argument is always a
string, and if errors are found in the API response, mw.Api will also
generate a rejection. This expands InstantApi to handle all of these
cases correctly.

Change-Id: I36a7ec85f7954302addd5a46fa988cdccfd83388