Fix type hint confusion around BatchedEntityTitleStoreLookup
This is a direct follow up for If4df542. Unfortunately there was much
more wrong with this code. Instead of a class that implements the
EntityTitleLookup interface, some of this code now expects a class that
implements *both* the EntityTitleLookup and the
BatchedEntityTitleStoreLookup interfaces.

Bot not all code! Some code is fine with one of the two.

Also no code actually expects a specific implementation. Type hints for
EntityContentFactory are not needed.

Change-Id: I18248909446b9342786d48150789a29cd4c27fd5