bridge: detect cascade-protected errors
Detect when the API indicates the page is page-protected, and map the
error accordingly in the CombiningPermissionsRepository. Note that we
had assumed this error would be exclusive to the repository, but in fact
it can happen on the client as well (wgIsProbablyEditable doesn’t seem
to take cascading protection into account), so an additional error type
has to be added to PageEditPermissionsRepository.

Bug: T239336
Change-Id: Ifc06782b7c63de16fb7ffb47a72c7fee7953adc7