bridge: Rename pseudo interface to be less confusing
This interface is actually not intended to be implemented anywhere but
servers as a map to be used in method signatures.

The alternative to using an interface would be using method overloading,
but that seems less maintainable.

However, if this is not a "real" interface, then the name should make
that somewhat transparent.

Change-Id: Icbc74817fd414b0851c9daa9a2c43649df6d9dfc