bridge: rename and clarify permissions interface
PageEditPermissionsRepository (checking repo + client) was confusingly
similar to PageEditPermissionErrorsRepository (checking one wiki). Let’s
rename it to hopefully better reflect its purpose.

Also, generalize the interface so it can be reused for different
item/page combinations, by passing the titles to the method rather than
the constructor. This is also more in line with other repositories we
have, including PageEditPermissionErrorsRepository. (The parameter names
and order – first repo, then client – are taken from the former
constructor of CombiningPermissionsRepository, though I renamed
repoPageTitle to repoItemTitle for additional clarity.)

Also, rename that method to make it more obvious that it’s not just for
one page, and include the parameter order (item then page) in the name.

Change-Id: I61e8d4be02b4ad6ee8e79fcd930957be902afd87