bridge: dispatch permission check on BRIDGE_INIT
Query repo and client if the user will be allowed to edit the item of
interest as seen on the page where bridge is loaded.

Since the end-to-end tests now end up making batched API requests, the
mocks are reorganized – the mock functions for the individual responses
all accept an argument with a partial response that they add their data
onto, and the mocked get() function always returns one aggregate object
that combines all the responses together. (Updating the mock functions
to update an incoming response seemed easier than writing a function to
merge individually-constructed responses after the fact.)

Bug: T241823
Change-Id: I0e0d3630ff2085c6c2748fde4128f3f0a22b4146