bridge: use reasonable defaults for mock services
Use an empty default services for mocks as defaults - this avoids
having to update seemingly unrelated places.
This also makes one unit test use mock services instead of a real
service container, making it less of an integration test which it was
before.
However, some "unit" tests (e.g. App.spec) still use a real store and
complex actions to set it into a desired state which in turn requires
many services to be mocked which have little to do with the subject
under test - to be improved.

Change-Id: I29e390fd9bc5e2b9bb4eb5add7bd2870ff524f8d