docs, Start documenting some key namespaces
This change sets up a system to add brief descriptions
to the auto generated documentation on doc.wikimedia.org

This also adds the @internal annotation to some doc blocks
that would otherwise be included in the namespace documentation.

This change also moves some test classes that were in
\Wikibase\Test to \Wikibase\Tests to be consistent with
other classes (eliminating 1 namespace)
There is still more cleanup to do regarding these namespaces,
as most of these tests are for repo maint scripts so perhaps
the test classes should be in \Wikibase\Repo\Tests.
but I will leave that for a followup.

Bug: T241215
Change-Id: I9f99650fe01ee6606e8d725bb1060827962945c9