docs, Add some Doxygen aliases (mostly from core)
I found this was needed while trying to hide the @licence
annotations from appearing in the class description summaries.

@licence is not a doxygen command, @copyright is however, so now
licence redirects to that as an alias.

I also copied a bunch of other aliases that we will end up using
from core.

The aliases that ='' (equal / redirect to a blank string)
mean that tags such as @covers, @codingStandardsIgnoreEnd etc
do not get included in things such as doxygen descriptions.

Bug: T241215
Change-Id: Ibc05a4e01d1439aab5944ddee778deca4d5e05a7