docs, overhaul 'options' docs
 - Split the existing options doc into linkable sections
 - Standardize the way examples and defaults are shown
 - Links to many mw.org pages when they are used for
   default settings
 - Link to new doc from other docs
 - Improve docs about id generation settings
 - Remove & update some comments in default php files
 - Add docs to Documentation.md about markdown and linking

It might make sense to split this very long .md file into
some sub files to make it easier to work with.

In the future we might want to use the doxygen group functionality
to move this documentation out of a markdown file and into the php
files that set the defaults.
However, not right now, and this change is already a big improvement.

Bug: T241215
Bug: T165973
Change-Id: I355c55efb4246e950e56a899562280a5c43b2402