doxygen, ignore pointless errors & fix others
Add some negative greps to ignore some errors that are
just spammy

Currently this only works when running doxygen locally
and not as part of the CI run.
But, this is a step in the right direction.

Also fix some other errors that we can now see in the
doxygen output.

PART 1:

Doxygen matches \foo as a command, and thus tries to turn
all php namespaces into commands, which is undesirable.

There is apparently no way to turn of \ and only use @
Thus, the next best thing is ignoring all of these
misstaken commands in the error ourput of doxygen
so that the remaining errors dont get flooded out.

PART 2:
@link annotation is used in php doc comments to provide in
IDE links, however, using these also causes doxygen to spout
errors as a full doxygen entity is not provided as the name.

So just hide these errors, as the behariour that ends up
resulting is actually desired

PART 3:
Ignore erorrs caused by @var annotations without declaring
the name of the var (only the type)

PART 4:
Ignore errors caused by {{#property}} parser function in docs
(and other similar parser functions) which doxygen tries to
parse as a link

Bug: T95116
Bug: T241215
Change-Id: I95365669fd4dc135edf22895cbbeab1c9aa899df