Simply return ids to insert instead of failing
The code that supposed to handle such cases is failing to consider all
sorts of valid usecases that can reach this path, causing hundreds of
fatals every day. This is also affecting other parts given that it
throws an exception basically turning the table. It can also cause
issues for other DeferredUpdates and for now it's better to just take
the existing ones into account. An edit on top of it would fix the issue
and given that it's secondary storage, 100% success is not needed or
possible.

Bug: T233414
Change-Id: I103805bba808588b69f0cb06806d8647a6661a15