wbterms: Add AliasTermBuffer
The current TermBuffer does not really account for aliases.
You can get a single alias via the various implementaitons
but not a whole list.
PageTerms is one example that requires a whole list.
PageTerms is currently backed by TermIndex (which is going away).

So this patch intorduces a new interface to add onto the old
one that allows retrieval of the prefecthed aliases in
implementations of the TermBuffer.
Within wikibase this is mainly PrefecthingTermLookup.

Once the entitysource migration and the wbterms migration
is done and there is generally less code lying around
this interface will need another pass and likely need moving
to data-model-services.

More work will likely want to be done with these interfaces
then too.

Bug: T242326
Depends-On: Ib012a1ad56330d920fc022db0c4ee1ba7dabcf01
Change-Id: I9019effdc3ee94020a2e604bfb7f25f8e87cca43