bridge: add router to build links to the repository
Adds a $repoRouter to the Vue prototype.

This mimicks parts of MediaWiki's `util.getUrl()`[0] but hard codes
the assumption that the repo's `wgScript` is 'index.php', as it is at
the moment not available from `RepoAccessModule` and seems to be hard
coded in few other places as well. For full configurability this could
be improved later.

On testing plugin registration see [1].

[0] https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/core/+/2e03bd7/resources/src/mediawiki.util/util.js#144
[1] https://github.com/eddyerburgh/vue-test-utils-jest-example/issues/6#issuecomment-350362860

Bug: T242413
Change-Id: I09d5e28ae6bbd214fe6daf92f0f22eaf9de70079