bridge: simplify ReferenceSection spacing
ReferenceSection used to have a more complicated spacing setup, where
the section itself began further to the left than visible, and both the
heading and the references were offset further to the right, apparently
intended to highlight a larger area on hover. Such highlighting on hover
has not yet been implemented, and it’s no longer clear from the mocks
that it’s supposed to be there, so for now we simplify the styles: the
whole reference section has a margin, and the heading and references
within it don’t have any further padding.

Extracted from I57801d9abeb8caa922f77763b259fedcd31cfb0a.

Change-Id: Id3e609b7061c2080b1814670c0c91b6277ab5a02
Co-authored-by: Michael Große <michael.grosse@wikimedia.de>