bridge: wire ErrorPermission into the ErrorWrapper
Show ErrorPermission inside of ErrorWrapper and pass it the permission
errors from the list of application errors.
This does not answer the challenge how we react if there are other
errors, in addition to the permission errors, inside of the application
errors.
This also removes the story for ErrorWrapper (called Error.stories.js)
as it is no longer a "leaf component" but is now store-dependent - will
move its visually interesting part into a dedicated component in a next
patch.

Bug: T237536
Change-Id: Iaa0e8a1f5f93e8c9d2c721bf693220cd5c9edbf8