bridge: format user names in blocked errors
Because jqueryMsg doesn’t parse wikitext in message parameters, in order
to make sure the user name doesn’t influence the surrounding text, we
need to pass an HTML element not just for $1 (user link: <a> + <bdi>)
but also for $4 (username: <bdi>), which means that $4 also can only be
used once in the message. (The alternative would be to surround the
username with Unicode First Strong Isolate (FSI) and Pop Directional
Isolate (PDI) characters, but those are not yet supported by all the
browsers we target [1][2].)

In ErrorPermission.spec.ts, the single parameterized test “interpolates
message with correct key and parameters” has been split up into multiple
tests. The test cases were already partially different (as evidenced by
the switch-case in repoRouterFactory), and splitting them up completely
simplifies them, in my opinion – parts that don’t matter to some test
cases can now be removed (e. g. several tests don’t need any router),
and where necessary, we can now tweak the assertions so that they still
work when HTML elements are passed into messages. (The tests for cascade
protection are similar enough that they are combined again, though in a
simpler version than the original.)

The explicit tests that the correct router is used for cascade
protection errors are dropped again, since that behavior is now already
checked by the other tests (by using unusedRouter() for routers that
should not be used).

[1]: https://w3c.github.io/i18n-tests/results/bidi-algorithm#rli_etc
[2]: https://bugs.webkit.org/show_bug.cgi?id=130823

Bug: T242832
Change-Id: I50812451a1b691b854a3d25add07f1c8e4f4d299