bridge: Hide save button on error states
This cause some "jumping" of the title when the app enters the error
state from the initializing state. However, this could only, if at all,
be helped in cases where both the heading and the save button label are
short and only with significant additional effort, which seems to be
beyond the scope of the behavior change.

Bug: T243010
Change-Id: I4d88ebbf356f3f392e9dfefce3ee181539360dbc