Overhaul TimeParser tests
These tests are very very very broken. PHPUnit never complained when
they tried to mock a non-existing class (I28258484a3df15af346aad17dfb4fa77ad43fa24), but in reality, these tests had several layers of brokenness. It ranges from private methods that should be public to infinite loop when using Language::factory in the service redefinition, with missing overrides in the middle, and dressed with redundant assertions and wrong test cases.

It's a shame that PHPUnit just bailed out immediately (and silently!)
instead of reporting any of those, since for humans it's pretty hard to
notice this kind of stupid mistakes when tests just pass.

Bug: T192167
Change-Id: I3ca2ccddc4ea985a3459d89ee9d29373d57dabc1