bridge: Create error types for bailout
This creates three different error types for the respective bailout
conditions.

The decision to use three types instead of one is to prevent the info
parameter being misused for some kind of subtyping instead of holding
supplementary information like a stack trace or data.

The validateBridgeApplicability has been refactored to no longer return
a value since it now can cause side effects with dispatching an action.

Bug: T242841
Change-Id: I4c28609870e19b3cdab83941bd40dd08e9269536