bridge: extract helper for creating API responses
The previous code to get or create a page in the response in the test
mocks was, I admit, pure laziness on my part. It worked fine since we
only wanted to mock a single response page anyways, and I could live
with it since it only happened in one function. But now that we have
more mock responses working with query pages, let’s create a helper
function that does the right thing, so that the mock functions can be
mixed more freely, without accidentally affecting the wrong pages.

Change-Id: I75dcba9257bcd64dc1c688a1a6faa421f94c3859