bridge: apply storybook resets in a scoped fashion
In production, by virtue of postcss-prefixwrap[0], all styles are scoped
to a shared selector (.wb-db-app). This is true for reset styles as
well, to avoid "unstyling" mediawiki through it.

In storybook the reset was applied in an unscoped fashion - affecting
storybook itself.

This moves the reset CSS into the "auto included" main.scss, avoiding
that it is accidentally included in a non-prod way, and ensures that
the production-like postcss-prefixwrap logic is applied in storybook,
too (making sure that the selector chains look the same).

Consequently production-like reset styles are applied inside of
storybook stories which should give us
* a more realistic behavior of components inside storybook
* avoids having to "reset resets" to make storybook itself look pretty
  (because elements outside the .wb-db-app selector never get unstyled)

This was brought to light by changes suggested in
I756d6599e0338864b2e82dc3c14ab5b0aff2bec1
but is no answer to it, as there the elements in question are part of
the story and consequently inside of the .wb-db-app selector.

[0] https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/Wikibase/+/3bd6e20/client/data-bridge/postcss.config.js#4

Change-Id: Ib6d12fe4d9a037e151357b74749db4a0768d9174