TR: Don't dispatch STOP_EDIT action on save
This patch only contains one non-refactoring change which
is removing the STOP_EDIT action on save.

This is not needed because wikibase.statement.stopEditing will
be called after save and not just on cancel.

It clarifies that in the naming of editing hooks in MWHookHandler

This patch adds a new test to cover the start editing hook. It also
contains refactoring to remove duplication in MWHookHandler tests.

Bug: T242119
Change-Id: I3e71f1d081951bb7a533b99e1afaf38f74250d19