bridge: rewrite store with vuex-smart-modules
This rewrites the store with the library
https://github.com/ktsn/vuex-smart-module to achieve more type safety
when committing mutations and dispatching actions. As a side effect the
store also got simpler.

The actions dispatched in the Vue components are not yet type safe. That
will be introduced in a follow-up patch.

Review note: start with the src/store/index.ts and go deeper from there.
Also, reference the README.md of vuex-smart-modules for more
information on the patterns used.

This structure is now focussed around modules and classes. The module
constructor gets the classes for Actions etc. themselves and constructs
them internally. This has the consequence that we cannot use their
constructors for dependency injection. The dependencies now have to be
injected via the magic $init method into getters and actions. We have to
add nested modules the same way in order to be able to use them.

There doesn't seem to be a good way to inject these dependencies and
nested modules in test. That is the reason that the tests use a lot of
// @ts-ignore annotations to silence typescript when we are overwriting
those private fields. I opened an issue at the library to ask for
clarification and documentation on how to do this properly, see
ktsn/vuex-smart-module#140

It seems to be not possible to nest modules deeper than one level, so
the statements module is now directly nested in the root module and no
longer in the entity module.

Change-Id: I8dea58b3b4c40b265d86b28b63cc8d692ac7929c