Cleanup repo Item vocab uri parsing & fix federated units
This mainly makes units work for the type of federation currently
deployed on commons (entity source based).

Before this change the local entity source was always fetched for
this vocab uri parsing, but on commons, the local source has no items.
So instead, select the item source...

This also renames a bunch of vars around the place to make sure that
in the future it is understood what exactly is being used further
down the chain.

Bug: T239474
Change-Id: I47d077cca2f8772db99874817b0d1f3c5b1c688f