bridge: Make center column side margins more consistent
This size of modal padding and sections margin is intended to be 40px
together. To remain aligned with OOUI, this leaves the padding at 16px
and adjusts the margin for both the right and left side to 24px. Both
are expressed as multiples of our base-unit 8px;

The variable for the base unit is renamed to make it more obvious that
this variable is intended to be a fix pixel value.  We may later want to
introduce another variable that depends on the browser font size.

Bug: T243584
Change-Id: I6dfebbfb5171edd2043961c26f5e56f571bfea05