bridge: make extended Window type explicit
Acknowledge that bridge relies on a window object extended by MediaWiki
in a prominent spot - when ever bridge makes use of this global object,
it is assumed that it has the appropriate type.
The MwWindow interface still makes sense to indicate this when passing
an object of this type around.

Change-Id: Ib4fec25a35ecc2e9f67dd655fdb331f4df6902b2