TR: Start Injecting tracking function into store
Introduced a new TrackFunction type that mirrors that provided
by the mediawiki global

Pass this to createStore which passes it down to the actions.
Currently it is unused there. A follup will see a proper trackFunction
injected and it being used.

To keep the tests green substanstial refactoring of tests for
components was needed.

Only App.vue test remains being tested with a complete store like
in production. Other components now only use a partial store where
the needed only parts are included.

Co-Authored-By: rosalie <perside.rosalie@wikimedia.de>

Bug: T242433
Change-Id: I2838c41f02585d9386a5527767c22460a10e71b5