wbterms: disallow using store methods when entities are non local
Follow up to I381e6b2826b90578cb5a83a8df25a220fca2c29c

Prior to this change the code just checked to see if the entity
source was defined and configured, which is will be fore remote
repos.
So add another check to see if the entity source it local or not
and only allow writing when local.

This change also adds a few bits of documentation and return types
to make following this specific code path a bit clearer.

This change also makes a fix to RepoHooks regarding entitysources
migration that I noticed due to the typehint that I added to
EntitySourceDefinitions::getSourceForEntityType
which would be a followup to Ib8046ff028ae3cf24d8af6facfb3182c439dc45a

Bug: T243769
Change-Id: I9dbfb95f6043423302c1bfcfb6a3271cf4c6416b